Revert "QuickAdminPanel file update"

+++ database/migrations/2020_10_13_000021_add_relationship_fields_to_departments_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToDepartmentsTable extends Migration
     public function up()
     {
         Schema::table('departments', function (Blueprint $table) {
-            $table->unsignedBigInteger('ministry_id')->nullable();
+            $table->unsignedInteger('ministry_id')->nullable();
             $table->foreign('ministry_id', 'ministry_fk_2164705')->references('id')->on('ministries');
         });
     }
