+++ database/migrations/2020_10_13_000023_add_relationship_fields_to_newsand_updates_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToNewsandUpdatesTable extends Migration
     public function up()
     {
         Schema::table('newsand_updates', function (Blueprint $table) {
-            $table->unsignedBigInteger('author_id')->nullable();
+            $table->unsignedInteger('author_id')->nullable();
             $table->foreign('author_id', 'author_fk_2180193')->references('id')->on('users');
         });
     }
