+++ database/migrations/2020_10_13_000025_add_relationship_fields_to_services_sub_categories_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToServicesSubCategoriesTable extends Migration
     public function up()
     {
         Schema::table('services_sub_categories', function (Blueprint $table) {
-            $table->unsignedBigInteger('servicescategory_id')->nullable();
+            $table->unsignedInteger('servicescategory_id')->nullable();
             $table->foreign('servicescategory_id', 'servicescategory_fk_2384426')->references('id')->on('service_categories');
         });
     }
